mImwzfqeMh/pyproject.toml
[project]
name = "weather-metrics-service"
version = "0.1.0"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.1",
    "SQLAlchemy==2.0.34",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.1",
    "pytest==8.3.3",
    "httpx==0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]